+++ source/Messaging/source/Communication/Subscriber/DeadLetterConstants.cs
@@ -1,0 +1,24 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+namespace Energinet.DataHub.Core.Messaging.Communication.Subscriber;
+
+public static class DeadLetterConstants
+{
+    /// <summary>
+    /// Suffix added to name of the original entity path (queue or subscription),
+    /// to get the full entity path to the dead-letter queue.
+    /// </summary>
+    public const string DeadLetterQueueSuffix = "/$DeadLetterQueue";
+}
