+++ source/Messaging/source/Communication/Subscriber/IDeadLetterHandler.cs
@@ -1,0 +1,29 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using Azure.Messaging.ServiceBus;
+using Microsoft.Azure.Functions.Worker;
+
+namespace Energinet.DataHub.Core.Messaging.Communication.Subscriber;
+
+/// <summary>
+/// Handler to support dead-letter queue handling in a function app.
+/// </summary>
+public interface IDeadLetterHandler
+{
+    /// <summary>
+    /// Handle a dead-lettered ServiceBus message.
+    /// </summary>
+    Task HandleAsync(ServiceBusReceivedMessage message, ServiceBusMessageActions messageActions);
+}
